Reduced complexity of box blur from O(r * r) per pixel to O(r) per pixel.

+++ src/main.h
@@ -60,7 +60,7 @@ typedef struct {
     double scaleStrict;
     bool merge;
     char* mergeFilePath;
-    uint32_t blur;
+    size_t blur;
     bool encode;
     char* encodeFilePath;
     bool experimental;
